refactor(postgrest): isolate and share types using symlink
