+++ packages/core/postgrest-js/src/PostgrestBuilder.ts
@@ -2,21 +2,20 @@
 import nodeFetch from '@supabase/node-fetch'
 
 import type {
-  Fetch,
   PostgrestSingleResponse,
   PostgrestResponseSuccess,
   CheckMatchingArrayTypes,
   MergePartialResult,
   IsValidResultOverride,
-  ClientServerOptions,
-} from './types'
+} from './types/types'
+import { ClientServerOptions, Fetch } from './types/common/common'
 import PostgrestError from './PostgrestError'
 import { ContainsNull } from './select-query-parser/types'
 
 export default abstract class PostgrestBuilder<
   ClientOptions extends ClientServerOptions,
   Result,
-  ThrowOnError extends boolean = false,
+  ThrowOnError extends boolean = false
 > implements
     PromiseLike<
       ThrowOnError extends true ? PostgrestResponseSuccess<Result> : PostgrestSingleResponse<Result>
@@ -85,7 +84,7 @@ export default abstract class PostgrestBuilder<
     TResult1 = ThrowOnError extends true
       ? PostgrestResponseSuccess<Result>
       : PostgrestSingleResponse<Result>,
-    TResult2 = never,
+    TResult2 = never
   >(
     onfulfilled?:
       | ((
@@ -275,7 +274,7 @@ export default abstract class PostgrestBuilder<
    */
   overrideTypes<
     NewResult,
-    Options extends { merge?: boolean } = { merge: true },
+    Options extends { merge?: boolean } = { merge: true }
   >(): PostgrestBuilder<
     ClientOptions,
     IsValidResultOverride<Result, NewResult, false, false> extends true
