+++ packages/core/postgrest-js/src/PostgrestQueryBuilder.ts
@@ -1,13 +1,19 @@
 import PostgrestFilterBuilder from './PostgrestFilterBuilder'
 import { GetResult } from './select-query-parser/result'
-import { ClientServerOptions, Fetch, GenericSchema, GenericTable, GenericView } from './types'
+import {
+  ClientServerOptions,
+  Fetch,
+  GenericSchema,
+  GenericTable,
+  GenericView,
+} from './types/common/common'
 
 export default class PostgrestQueryBuilder<
   ClientOptions extends ClientServerOptions,
   Schema extends GenericSchema,
   Relation extends GenericTable | GenericView,
   RelationName = unknown,
-  Relationships = Relation extends { Relationships: infer R } ? R : unknown,
+  Relationships = Relation extends { Relationships: infer R } ? R : unknown
 > {
   url: URL
   headers: Headers
@@ -63,7 +69,7 @@ export default class PostgrestQueryBuilder<
       Relationships,
       Query,
       ClientOptions
-    >,
+    >
   >(
     columns?: Query,
     options?: {
