+++ packages/core/postgrest-js/src/select-query-parser/types.ts
@@ -1,6 +1,22 @@
-import type { GenericRelationship, GenericSchema, GenericTable, Prettify } from '../types'
+import type {
+  GenericRelationship,
+  GenericSchema,
+  GenericTable,
+  ClientServerOptions,
+  GenericSetofOption,
+  GenericFunction,
+} from '../types/common/common'
+import type { Prettify } from '../types/types'
 
-export type { GenericRelationship, GenericSchema, GenericTable, Prettify }
+export type {
+  GenericRelationship,
+  GenericSchema,
+  GenericTable,
+  ClientServerOptions,
+  GenericSetofOption,
+  Prettify,
+  GenericFunction,
+}
 
 export type AggregateWithoutColumnFunctions = 'count'
 
@@ -54,16 +70,16 @@ type ArrayPostgreSQLTypes = `_${SingleValuePostgreSQLTypes}`
 type TypeScriptSingleValueTypes<T extends SingleValuePostgreSQLTypes> = T extends 'bool'
   ? boolean
   : T extends PostgresSQLNumberTypes
-    ? number
-    : T extends PostgresSQLStringTypes
-      ? string
-      : T extends 'json' | 'jsonb'
-        ? Json
-        : T extends 'void'
-          ? undefined
-          : T extends 'record'
-            ? Record<string, unknown>
-            : unknown
+  ? number
+  : T extends PostgresSQLStringTypes
+  ? string
+  : T extends 'json' | 'jsonb'
+  ? Json
+  : T extends 'void'
+  ? undefined
+  : T extends 'record'
+  ? Record<string, unknown>
+  : unknown
 
 type StripUnderscore<T extends string> = T extends `_${infer U}` ? U : T
 
@@ -82,8 +98,9 @@ export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) ex
   ? I
   : never
 
-export type LastOf<T> =
-  UnionToIntersection<T extends any ? () => T : never> extends () => infer R ? R : never
+export type LastOf<T> = UnionToIntersection<T extends any ? () => T : never> extends () => infer R
+  ? R
+  : never
 
 export type Push<T extends any[], V> = [...T, V]
 
@@ -100,8 +117,9 @@ export type ExtractFirstProperty<T> = T extends { [K in keyof T]: infer U } ? U 
 // Type predicates
 export type ContainsNull<T> = null extends T ? true : false
 
-export type IsNonEmptyArray<T> =
-  Exclude<T, undefined> extends readonly [unknown, ...unknown[]] ? true : false
+export type IsNonEmptyArray<T> = Exclude<T, undefined> extends readonly [unknown, ...unknown[]]
+  ? true
+  : false
 
 // Types for working with database schemas
 export type TablesAndViews<Schema extends GenericSchema> = Schema['Tables'] &
@@ -109,5 +127,5 @@ export type TablesAndViews<Schema extends GenericSchema> = Schema['Tables'] &
 
 export type GetTableRelationships<
   Schema extends GenericSchema,
-  Tname extends string,
+  Tname extends string
 > = TablesAndViews<Schema>[Tname] extends { Relationships: infer R } ? R : false
