+++ packages/core/postgrest-js/src/select-query-parser/utils.ts
@@ -1,4 +1,3 @@
-import { GenericFunction, GenericSetofOption } from '../types'
 import { Ast } from './parser'
 import {
   AggregateFunctions,
@@ -9,6 +8,8 @@ import {
   IsNonEmptyArray,
   TablesAndViews,
   UnionToArray,
+  GenericFunction,
+  GenericSetofOption,
 } from './types'
 
 export type IsAny<T> = 0 extends 1 & T ? true : false
@@ -25,7 +26,7 @@ export type SelectQueryError<Message extends string> = { error: true } & Message
  */
 export type DeduplicateRelationships<T extends readonly unknown[]> = T extends readonly [
   infer First,
-  ...infer Rest,
+  ...infer Rest
 ]
   ? First extends Rest[number]
     ? DeduplicateRelationships<Rest extends readonly unknown[] ? Rest : []>
@@ -35,18 +36,18 @@ export type DeduplicateRelationships<T extends readonly unknown[]> = T extends r
 export type GetFieldNodeResultName<Field extends Ast.FieldNode> = Field['alias'] extends string
   ? Field['alias']
   : Field['aggregateFunction'] extends AggregateFunctions
-    ? Field['aggregateFunction']
-    : Field['name']
+  ? Field['aggregateFunction']
+  : Field['name']
 
 type FilterRelationNodes<Nodes extends Ast.Node[]> = UnionToArray<
   {
     [K in keyof Nodes]: Nodes[K] extends Ast.SpreadNode
       ? Nodes[K]['target']
       : Nodes[K] extends Ast.FieldNode
-        ? IsNonEmptyArray<Nodes[K]['children']> extends true
-          ? Nodes[K]
-          : never
+      ? IsNonEmptyArray<Nodes[K]['children']> extends true
+        ? Nodes[K]
         : never
+      : never
   }[number]
 >
 
@@ -54,7 +55,7 @@ type ResolveRelationships<
   Schema extends GenericSchema,
   RelationName extends string,
   Relationships extends GenericRelationship[],
-  Nodes extends Ast.FieldNode[],
+  Nodes extends Ast.FieldNode[]
 > = UnionToArray<{
   [K in keyof Nodes]: Nodes[K] extends Ast.FieldNode
     ? ResolveRelationship<Schema, Relationships, Nodes[K], RelationName> extends infer Relation
@@ -117,32 +118,31 @@ export type CheckDuplicateEmbededReference<
   Schema extends GenericSchema,
   RelationName extends string,
   Relationships extends GenericRelationship[],
-  Nodes extends Ast.Node[],
-> =
-  FilterRelationNodes<Nodes> extends infer RelationsNodes
-    ? RelationsNodes extends Ast.FieldNode[]
-      ? ResolveRelationships<
-          Schema,
-          RelationName,
-          Relationships,
-          RelationsNodes
-        > extends infer ResolvedRels
-        ? ResolvedRels extends unknown[]
-          ? FindDuplicates<ResolvedRels> extends infer Duplicates
-            ? Duplicates extends never
-              ? false
-              : Duplicates extends { fieldName: infer FieldName }
-                ? FieldName extends string
-                  ? {
-                      [K in FieldName]: SelectQueryError<`table "${RelationName}" specified more than once use hinting for desambiguation`>
-                    }
-                  : false
-                : false
+  Nodes extends Ast.Node[]
+> = FilterRelationNodes<Nodes> extends infer RelationsNodes
+  ? RelationsNodes extends Ast.FieldNode[]
+    ? ResolveRelationships<
+        Schema,
+        RelationName,
+        Relationships,
+        RelationsNodes
+      > extends infer ResolvedRels
+      ? ResolvedRels extends unknown[]
+        ? FindDuplicates<ResolvedRels> extends infer Duplicates
+          ? Duplicates extends never
+            ? false
+            : Duplicates extends { fieldName: infer FieldName }
+            ? FieldName extends string
+              ? {
+                  [K in FieldName]: SelectQueryError<`table "${RelationName}" specified more than once use hinting for desambiguation`>
+                }
+              : false
             : false
           : false
         : false
       : false
     : false
+  : false
 
 /**
  * Returns a boolean representing whether there is a foreign key referencing
@@ -153,16 +153,16 @@ type HasFKeyToFRel<FRelName, Relationships> = Relationships extends [infer R]
     ? true
     : false
   : Relationships extends [infer R, ...infer Rest]
-    ? HasFKeyToFRel<FRelName, [R]> extends true
-      ? true
-      : HasFKeyToFRel<FRelName, Rest>
-    : false
+  ? HasFKeyToFRel<FRelName, [R]> extends true
+    ? true
+    : HasFKeyToFRel<FRelName, Rest>
+  : false
 /**
  * Checks if there is more than one relation to a given foreign relation name in the Relationships.
  */
 type HasMultipleFKeysToFRelDeduplicated<FRelName, Relationships> = Relationships extends [
   infer R,
-  ...infer Rest,
+  ...infer Rest
 ]
   ? R extends { referencedRelation: FRelName }
     ? HasFKeyToFRel<FRelName, Rest> extends true
@@ -173,52 +173,51 @@ type HasMultipleFKeysToFRelDeduplicated<FRelName, Relationships> = Relationships
 
 type HasMultipleFKeysToFRel<
   FRelName,
-  Relationships extends unknown[],
+  Relationships extends unknown[]
 > = HasMultipleFKeysToFRelDeduplicated<FRelName, DeduplicateRelationships<Relationships>>
 
 type CheckRelationshipError<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
   CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-  FoundRelation,
-> =
-  FoundRelation extends SelectQueryError<string>
-    ? FoundRelation
-    : // If the relation is a reverse relation with no hint (matching by name)
-      FoundRelation extends {
-          relation: {
-            referencedRelation: infer RelatedRelationName
-            name: string
-          }
-          direction: 'reverse'
-        }
-      ? RelatedRelationName extends string
-        ? // We check if there is possible confusion with other relations with this table
-          HasMultipleFKeysToFRel<RelatedRelationName, Relationships> extends true
-          ? // If there is, postgrest will fail at runtime, and require desambiguation via hinting
-            SelectQueryError<`Could not embed because more than one relationship was found for '${RelatedRelationName}' and '${CurrentTableOrView}' you need to hint the column with ${RelatedRelationName}!<columnName> ?`>
-          : FoundRelation
-        : never
-      : // Same check for forward relationships, but we must gather the relationships from the found relation
-        FoundRelation extends {
-            relation: {
-              referencedRelation: infer RelatedRelationName
-              name: string
-            }
-            direction: 'forward'
-            from: infer From
-          }
-        ? RelatedRelationName extends string
-          ? From extends keyof TablesAndViews<Schema> & string
-            ? HasMultipleFKeysToFRel<
-                RelatedRelationName,
-                TablesAndViews<Schema>[From]['Relationships']
-              > extends true
-              ? SelectQueryError<`Could not embed because more than one relationship was found for '${From}' and '${RelatedRelationName}' you need to hint the column with ${From}!<columnName> ?`>
-              : FoundRelation
-            : never
-          : never
+  FoundRelation
+> = FoundRelation extends SelectQueryError<string>
+  ? FoundRelation
+  : // If the relation is a reverse relation with no hint (matching by name)
+  FoundRelation extends {
+      relation: {
+        referencedRelation: infer RelatedRelationName
+        name: string
+      }
+      direction: 'reverse'
+    }
+  ? RelatedRelationName extends string
+    ? // We check if there is possible confusion with other relations with this table
+      HasMultipleFKeysToFRel<RelatedRelationName, Relationships> extends true
+      ? // If there is, postgrest will fail at runtime, and require desambiguation via hinting
+        SelectQueryError<`Could not embed because more than one relationship was found for '${RelatedRelationName}' and '${CurrentTableOrView}' you need to hint the column with ${RelatedRelationName}!<columnName> ?`>
+      : FoundRelation
+    : never
+  : // Same check for forward relationships, but we must gather the relationships from the found relation
+  FoundRelation extends {
+      relation: {
+        referencedRelation: infer RelatedRelationName
+        name: string
+      }
+      direction: 'forward'
+      from: infer From
+    }
+  ? RelatedRelationName extends string
+    ? From extends keyof TablesAndViews<Schema> & string
+      ? HasMultipleFKeysToFRel<
+          RelatedRelationName,
+          TablesAndViews<Schema>[From]['Relationships']
+        > extends true
+        ? SelectQueryError<`Could not embed because more than one relationship was found for '${From}' and '${RelatedRelationName}' you need to hint the column with ${From}!<columnName> ?`>
         : FoundRelation
+      : never
+    : never
+  : FoundRelation
 /**
  * Resolves relationships for embedded resources and retrieves the referenced Table
  */
@@ -226,23 +225,22 @@ export type ResolveRelationship<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
   Field extends Ast.FieldNode,
-  CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-> =
-  ResolveReverseRelationship<
-    Schema,
-    Relationships,
-    Field,
-    CurrentTableOrView
-  > extends infer ReverseRelationship
-    ? ReverseRelationship extends false
-      ? CheckRelationshipError<
-          Schema,
-          Relationships,
-          CurrentTableOrView,
-          ResolveForwardRelationship<Schema, Field, CurrentTableOrView>
-        >
-      : CheckRelationshipError<Schema, Relationships, CurrentTableOrView, ReverseRelationship>
-    : never
+  CurrentTableOrView extends keyof TablesAndViews<Schema> & string
+> = ResolveReverseRelationship<
+  Schema,
+  Relationships,
+  Field,
+  CurrentTableOrView
+> extends infer ReverseRelationship
+  ? ReverseRelationship extends false
+    ? CheckRelationshipError<
+        Schema,
+        Relationships,
+        CurrentTableOrView,
+        ResolveForwardRelationship<Schema, Field, CurrentTableOrView>
+      >
+    : CheckRelationshipError<Schema, Relationships, CurrentTableOrView, ReverseRelationship>
+  : never
 
 /**
  * Resolves reverse relationships (from children to parent)
@@ -251,40 +249,39 @@ type ResolveReverseRelationship<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
   Field extends Ast.FieldNode,
-  CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-> =
-  FindFieldMatchingRelationships<Schema, Relationships, Field> extends infer FoundRelation
-    ? FoundRelation extends never
-      ? false
-      : FoundRelation extends { referencedRelation: infer RelatedRelationName }
-        ? RelatedRelationName extends string
-          ? RelatedRelationName extends keyof TablesAndViews<Schema>
-            ? // If the relation was found via hinting we just return it without any more checks
-              FoundRelation extends { hint: string }
-              ? {
-                  referencedTable: TablesAndViews<Schema>[RelatedRelationName]
-                  relation: FoundRelation
-                  direction: 'reverse'
-                  from: CurrentTableOrView
-                }
-              : // If the relation was found via implicit relation naming, we must ensure there is no conflicting matches
-                HasMultipleFKeysToFRel<RelatedRelationName, Relationships> extends true
-                ? SelectQueryError<`Could not embed because more than one relationship was found for '${RelatedRelationName}' and '${CurrentTableOrView}' you need to hint the column with ${RelatedRelationName}!<columnName> ?`>
-                : {
-                    referencedTable: TablesAndViews<Schema>[RelatedRelationName]
-                    relation: FoundRelation
-                    direction: 'reverse'
-                    from: CurrentTableOrView
-                  }
-            : SelectQueryError<`Relation '${RelatedRelationName}' not found in schema.`>
-          : false
-        : false
+  CurrentTableOrView extends keyof TablesAndViews<Schema> & string
+> = FindFieldMatchingRelationships<Schema, Relationships, Field> extends infer FoundRelation
+  ? FoundRelation extends never
+    ? false
+    : FoundRelation extends { referencedRelation: infer RelatedRelationName }
+    ? RelatedRelationName extends string
+      ? RelatedRelationName extends keyof TablesAndViews<Schema>
+        ? // If the relation was found via hinting we just return it without any more checks
+          FoundRelation extends { hint: string }
+          ? {
+              referencedTable: TablesAndViews<Schema>[RelatedRelationName]
+              relation: FoundRelation
+              direction: 'reverse'
+              from: CurrentTableOrView
+            }
+          : // If the relation was found via implicit relation naming, we must ensure there is no conflicting matches
+          HasMultipleFKeysToFRel<RelatedRelationName, Relationships> extends true
+          ? SelectQueryError<`Could not embed because more than one relationship was found for '${RelatedRelationName}' and '${CurrentTableOrView}' you need to hint the column with ${RelatedRelationName}!<columnName> ?`>
+          : {
+              referencedTable: TablesAndViews<Schema>[RelatedRelationName]
+              relation: FoundRelation
+              direction: 'reverse'
+              from: CurrentTableOrView
+            }
+        : SelectQueryError<`Relation '${RelatedRelationName}' not found in schema.`>
+      : false
     : false
+  : false
 
 export type FindMatchingTableRelationships<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
-  value extends string,
+  value extends string
 > = Relationships extends [infer R, ...infer Rest]
   ? Rest extends GenericRelationship[]
     ? R extends { referencedRelation: infer ReferencedRelation }
@@ -292,10 +289,10 @@ export type FindMatchingTableRelationships<
         ? R extends { foreignKeyName: value }
           ? R & { match: 'fkname' }
           : R extends { referencedRelation: value }
-            ? R & { match: 'refrel' }
-            : R extends { columns: [value] }
-              ? R & { match: 'col' }
-              : FindMatchingTableRelationships<Schema, Rest, value>
+          ? R & { match: 'refrel' }
+          : R extends { columns: [value] }
+          ? R & { match: 'col' }
+          : FindMatchingTableRelationships<Schema, Rest, value>
         : FindMatchingTableRelationships<Schema, Rest, value>
       : false
     : false
@@ -304,7 +301,7 @@ export type FindMatchingTableRelationships<
 export type FindMatchingViewRelationships<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
-  value extends string,
+  value extends string
 > = Relationships extends [infer R, ...infer Rest]
   ? Rest extends GenericRelationship[]
     ? R extends { referencedRelation: infer ReferencedRelation }
@@ -312,10 +309,10 @@ export type FindMatchingViewRelationships<
         ? R extends { foreignKeyName: value }
           ? R & { match: 'fkname' }
           : R extends { referencedRelation: value }
-            ? R & { match: 'refrel' }
-            : R extends { columns: [value] }
-              ? R & { match: 'col' }
-              : FindMatchingViewRelationships<Schema, Rest, value>
+          ? R & { match: 'refrel' }
+          : R extends { columns: [value] }
+          ? R & { match: 'col' }
+          : FindMatchingViewRelationships<Schema, Rest, value>
         : FindMatchingViewRelationships<Schema, Rest, value>
       : false
     : false
@@ -325,7 +322,7 @@ export type FindMatchingHintTableRelationships<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
   hint extends string,
-  name extends string,
+  name extends string
 > = Relationships extends [infer R, ...infer Rest]
   ? Rest extends GenericRelationship[]
     ? R extends { referencedRelation: infer ReferencedRelation }
@@ -333,10 +330,10 @@ export type FindMatchingHintTableRelationships<
         ? R extends { foreignKeyName: hint }
           ? R & { match: 'fkname' }
           : R extends { referencedRelation: hint }
-            ? R & { match: 'refrel' }
-            : R extends { columns: [hint] }
-              ? R & { match: 'col' }
-              : FindMatchingHintTableRelationships<Schema, Rest, hint, name>
+          ? R & { match: 'refrel' }
+          : R extends { columns: [hint] }
+          ? R & { match: 'col' }
+          : FindMatchingHintTableRelationships<Schema, Rest, hint, name>
         : FindMatchingHintTableRelationships<Schema, Rest, hint, name>
       : false
     : false
@@ -345,7 +342,7 @@ export type FindMatchingHintViewRelationships<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
   hint extends string,
-  name extends string,
+  name extends string
 > = Relationships extends [infer R, ...infer Rest]
   ? Rest extends GenericRelationship[]
     ? R extends { referencedRelation: infer ReferencedRelation }
@@ -353,10 +350,10 @@ export type FindMatchingHintViewRelationships<
         ? R extends { foreignKeyName: hint }
           ? R & { match: 'fkname' }
           : R extends { referencedRelation: hint }
-            ? R & { match: 'refrel' }
-            : R extends { columns: [hint] }
-              ? R & { match: 'col' }
-              : FindMatchingHintViewRelationships<Schema, Rest, hint, name>
+          ? R & { match: 'refrel' }
+          : R extends { columns: [hint] }
+          ? R & { match: 'col' }
+          : FindMatchingHintViewRelationships<Schema, Rest, hint, name>
         : FindMatchingHintViewRelationships<Schema, Rest, hint, name>
       : false
     : false
@@ -364,7 +361,7 @@ export type FindMatchingHintViewRelationships<
 
 type IsColumnsNullable<
   Table extends Pick<GenericTable, 'Row'>,
-  Columns extends (keyof Table['Row'])[],
+  Columns extends (keyof Table['Row'])[]
 > = Columns extends [infer Column, ...infer Rest]
   ? Column extends keyof Table['Row']
     ? ContainsNull<Table['Row'][Column]> extends true
@@ -376,12 +373,12 @@ type IsColumnsNullable<
 // Check weither or not a 1-1 relation is nullable by checking against the type of the columns
 export type IsRelationNullable<
   Table extends GenericTable,
-  Relation extends GenericRelationship,
+  Relation extends GenericRelationship
 > = IsColumnsNullable<Table, Relation['columns']>
 
 type TableForwardRelationships<
   Schema extends GenericSchema,
-  TName,
+  TName
 > = TName extends keyof TablesAndViews<Schema>
   ? UnionToArray<
       RecursivelyFindRelationships<Schema, TName, keyof TablesAndViews<Schema>>
@@ -395,7 +392,7 @@ type TableForwardRelationships<
 type RecursivelyFindRelationships<
   Schema extends GenericSchema,
   TName,
-  Keys extends keyof TablesAndViews<Schema>,
+  Keys extends keyof TablesAndViews<Schema>
 > = Keys extends infer K
   ? K extends keyof TablesAndViews<Schema>
     ? FilterRelationships<TablesAndViews<Schema>[K]['Relationships'], TName, K> extends never
@@ -415,83 +412,82 @@ type FilterRelationships<R, TName, From> = R extends readonly (infer Rel)[]
 export type ResolveForwardRelationship<
   Schema extends GenericSchema,
   Field extends Ast.FieldNode,
-  CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-> =
-  FindFieldMatchingRelationships<
-    Schema,
-    TablesAndViews<Schema>[Field['name']]['Relationships'],
-    Ast.FieldNode & { name: CurrentTableOrView; hint: Field['hint'] }
-  > extends infer FoundByName
-    ? FoundByName extends GenericRelationship
+  CurrentTableOrView extends keyof TablesAndViews<Schema> & string
+> = FindFieldMatchingRelationships<
+  Schema,
+  TablesAndViews<Schema>[Field['name']]['Relationships'],
+  Ast.FieldNode & { name: CurrentTableOrView; hint: Field['hint'] }
+> extends infer FoundByName
+  ? FoundByName extends GenericRelationship
+    ? {
+        referencedTable: TablesAndViews<Schema>[Field['name']]
+        relation: FoundByName
+        direction: 'forward'
+        from: Field['name']
+        type: 'found-by-name'
+      }
+    : FindFieldMatchingRelationships<
+        Schema,
+        TableForwardRelationships<Schema, CurrentTableOrView>,
+        Field
+      > extends infer FoundByMatch
+    ? FoundByMatch extends GenericRelationship & {
+        from: keyof TablesAndViews<Schema>
+      }
       ? {
-          referencedTable: TablesAndViews<Schema>[Field['name']]
-          relation: FoundByName
+          referencedTable: TablesAndViews<Schema>[FoundByMatch['from']]
+          relation: FoundByMatch
           direction: 'forward'
-          from: Field['name']
-          type: 'found-by-name'
+          from: CurrentTableOrView
+          type: 'found-by-match'
         }
-      : FindFieldMatchingRelationships<
-            Schema,
-            TableForwardRelationships<Schema, CurrentTableOrView>,
-            Field
-          > extends infer FoundByMatch
-        ? FoundByMatch extends GenericRelationship & {
-            from: keyof TablesAndViews<Schema>
+      : FindJoinTableRelationship<
+          Schema,
+          CurrentTableOrView,
+          Field['name']
+        > extends infer FoundByJoinTable
+      ? FoundByJoinTable extends GenericRelationship
+        ? {
+            referencedTable: TablesAndViews<Schema>[FoundByJoinTable['referencedRelation']]
+            relation: FoundByJoinTable & { match: 'refrel' }
+            direction: 'forward'
+            from: CurrentTableOrView
+            type: 'found-by-join-table'
           }
+        : ResolveEmbededFunctionJoinTableRelationship<
+            Schema,
+            CurrentTableOrView,
+            Field['name']
+          > extends infer FoundEmbededFunctionJoinTableRelation
+        ? FoundEmbededFunctionJoinTableRelation extends GenericSetofOption
           ? {
-              referencedTable: TablesAndViews<Schema>[FoundByMatch['from']]
-              relation: FoundByMatch
+              referencedTable: TablesAndViews<Schema>[FoundEmbededFunctionJoinTableRelation['to']]
+              relation: {
+                foreignKeyName: `${Field['name']}_${CurrentTableOrView}_${FoundEmbededFunctionJoinTableRelation['to']}_forward`
+                columns: []
+                isOneToOne: FoundEmbededFunctionJoinTableRelation['isOneToOne'] extends true
+                  ? true
+                  : false
+                referencedColumns: []
+                referencedRelation: FoundEmbededFunctionJoinTableRelation['to']
+              } & {
+                match: 'func'
+                isNotNullable: FoundEmbededFunctionJoinTableRelation['isNotNullable'] extends true
+                  ? true
+                  : FoundEmbededFunctionJoinTableRelation['isSetofReturn'] extends true
+                  ? false
+                  : true
+                isSetofReturn: FoundEmbededFunctionJoinTableRelation['isSetofReturn']
+              }
               direction: 'forward'
               from: CurrentTableOrView
-              type: 'found-by-match'
+              type: 'found-by-embeded-function'
             }
-          : FindJoinTableRelationship<
-                Schema,
-                CurrentTableOrView,
-                Field['name']
-              > extends infer FoundByJoinTable
-            ? FoundByJoinTable extends GenericRelationship
-              ? {
-                  referencedTable: TablesAndViews<Schema>[FoundByJoinTable['referencedRelation']]
-                  relation: FoundByJoinTable & { match: 'refrel' }
-                  direction: 'forward'
-                  from: CurrentTableOrView
-                  type: 'found-by-join-table'
-                }
-              : ResolveEmbededFunctionJoinTableRelationship<
-                    Schema,
-                    CurrentTableOrView,
-                    Field['name']
-                  > extends infer FoundEmbededFunctionJoinTableRelation
-                ? FoundEmbededFunctionJoinTableRelation extends GenericSetofOption
-                  ? {
-                      referencedTable: TablesAndViews<Schema>[FoundEmbededFunctionJoinTableRelation['to']]
-                      relation: {
-                        foreignKeyName: `${Field['name']}_${CurrentTableOrView}_${FoundEmbededFunctionJoinTableRelation['to']}_forward`
-                        columns: []
-                        isOneToOne: FoundEmbededFunctionJoinTableRelation['isOneToOne'] extends true
-                          ? true
-                          : false
-                        referencedColumns: []
-                        referencedRelation: FoundEmbededFunctionJoinTableRelation['to']
-                      } & {
-                        match: 'func'
-                        isNotNullable: FoundEmbededFunctionJoinTableRelation['isNotNullable'] extends true
-                          ? true
-                          : FoundEmbededFunctionJoinTableRelation['isSetofReturn'] extends true
-                            ? false
-                            : true
-                        isSetofReturn: FoundEmbededFunctionJoinTableRelation['isSetofReturn']
-                      }
-                      direction: 'forward'
-                      from: CurrentTableOrView
-                      type: 'found-by-embeded-function'
-                    }
-                  : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
-                : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
-            : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
+          : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
         : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
+      : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
     : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
+  : SelectQueryError<`could not find the relation between ${CurrentTableOrView} and ${Field['name']}`>
 
 /**
  * Given a CurrentTableOrView, finds all join tables to this relation.
@@ -514,7 +510,7 @@ export type ResolveForwardRelationship<
 type ResolveJoinTableRelationship<
   Schema extends GenericSchema,
   CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-  FieldName extends string,
+  FieldName extends string
 > = {
   [TableName in keyof TablesAndViews<Schema>]: DeduplicateRelationships<
     TablesAndViews<Schema>[TableName]['Relationships']
@@ -534,34 +530,32 @@ type ResolveJoinTableRelationship<
 type ResolveEmbededFunctionJoinTableRelationship<
   Schema extends GenericSchema,
   CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-  FieldName extends string,
-> =
-  FindMatchingFunctionBySetofFrom<
-    Schema['Functions'][FieldName],
-    CurrentTableOrView
-  > extends infer Fn
-    ? Fn extends GenericFunction
-      ? Fn['SetofOptions']
-      : false
+  FieldName extends string
+> = FindMatchingFunctionBySetofFrom<
+  Schema['Functions'][FieldName],
+  CurrentTableOrView
+> extends infer Fn
+  ? Fn extends GenericFunction
+    ? Fn['SetofOptions']
     : false
+  : false
 
 export type FindJoinTableRelationship<
   Schema extends GenericSchema,
   CurrentTableOrView extends keyof TablesAndViews<Schema> & string,
-  FieldName extends string,
-> =
-  ResolveJoinTableRelationship<Schema, CurrentTableOrView, FieldName> extends infer Result
-    ? [Result] extends [never]
-      ? false
-      : Result
-    : never
+  FieldName extends string
+> = ResolveJoinTableRelationship<Schema, CurrentTableOrView, FieldName> extends infer Result
+  ? [Result] extends [never]
+    ? false
+    : Result
+  : never
 /**
  * Finds a matching relationship based on the FieldNode's name and optional hint.
  */
 export type FindFieldMatchingRelationships<
   Schema extends GenericSchema,
   Relationships extends GenericRelationship[],
-  Field extends Ast.FieldNode,
+  Field extends Ast.FieldNode
 > = Field extends { hint: string }
   ? FindMatchingHintTableRelationships<
       Schema,
@@ -574,97 +568,65 @@ export type FindFieldMatchingRelationships<
         hint: Field['hint']
       }
     : FindMatchingHintViewRelationships<
-          Schema,
-          Relationships,
-          Field['hint'],
-          Field['name']
-        > extends GenericRelationship
-      ? FindMatchingHintViewRelationships<Schema, Relationships, Field['hint'], Field['name']> & {
-          branch: 'found-in-view-via-hint'
-          hint: Field['hint']
-        }
-      : SelectQueryError<'Failed to find matching relation via hint'>
-  : FindMatchingTableRelationships<Schema, Relationships, Field['name']> extends GenericRelationship
-    ? FindMatchingTableRelationships<Schema, Relationships, Field['name']> & {
-        branch: 'found-in-table-via-name'
-        name: Field['name']
+        Schema,
+        Relationships,
+        Field['hint'],
+        Field['name']
+      > extends GenericRelationship
+    ? FindMatchingHintViewRelationships<Schema, Relationships, Field['hint'], Field['name']> & {
+        branch: 'found-in-view-via-hint'
+        hint: Field['hint']
       }
-    : FindMatchingViewRelationships<
-          Schema,
-          Relationships,
-          Field['name']
-        > extends GenericRelationship
-      ? FindMatchingViewRelationships<Schema, Relationships, Field['name']> & {
-          branch: 'found-in-view-via-name'
-          name: Field['name']
-        }
-      : SelectQueryError<'Failed to find matching relation via name'>
+    : SelectQueryError<'Failed to find matching relation via hint'>
+  : FindMatchingTableRelationships<Schema, Relationships, Field['name']> extends GenericRelationship
+  ? FindMatchingTableRelationships<Schema, Relationships, Field['name']> & {
+      branch: 'found-in-table-via-name'
+      name: Field['name']
+    }
+  : FindMatchingViewRelationships<Schema, Relationships, Field['name']> extends GenericRelationship
+  ? FindMatchingViewRelationships<Schema, Relationships, Field['name']> & {
+      branch: 'found-in-view-via-name'
+      name: Field['name']
+    }
+  : SelectQueryError<'Failed to find matching relation via name'>
 
 export type JsonPathToAccessor<Path extends string> = Path extends `${infer P1}->${infer P2}`
   ? P2 extends `>${infer Rest}` // Handle ->> operator
     ? JsonPathToAccessor<`${P1}.${Rest}`>
     : P2 extends string // Handle -> operator
-      ? JsonPathToAccessor<`${P1}.${P2}`>
-      : Path
+    ? JsonPathToAccessor<`${P1}.${P2}`>
+    : Path
   : Path extends `>${infer Rest}` // Clean up any remaining > characters
-    ? JsonPathToAccessor<Rest>
-    : Path extends `${infer P1}::${infer _}` // Handle type casting
-      ? JsonPathToAccessor<P1>
-      : Path extends `${infer P1}${')' | ','}${infer _}` // Handle closing parenthesis and comma
-        ? P1
-        : Path
+  ? JsonPathToAccessor<Rest>
+  : Path extends `${infer P1}::${infer _}` // Handle type casting
+  ? JsonPathToAccessor<P1>
+  : Path extends `${infer P1}${')' | ','}${infer _}` // Handle closing parenthesis and comma
+  ? P1
+  : Path
 
 export type JsonPathToType<T, Path extends string> = Path extends ''
   ? T
   : ContainsNull<T> extends true
-    ? JsonPathToType<Exclude<T, null>, Path>
-    : Path extends `${infer Key}.${infer Rest}`
-      ? Key extends keyof T
-        ? JsonPathToType<T[Key], Rest>
-        : never
-      : Path extends keyof T
-        ? T[Path]
-        : never
+  ? JsonPathToType<Exclude<T, null>, Path>
+  : Path extends `${infer Key}.${infer Rest}`
+  ? Key extends keyof T
+    ? JsonPathToType<T[Key], Rest>
+    : never
+  : Path extends keyof T
+  ? T[Path]
+  : never
 
 export type IsStringUnion<T> = string extends T
   ? false
   : T extends string
-    ? [T] extends [never]
-      ? false
-      : true
-    : false
-
-// Functions matching utils
-type IsMatchingArgs<
-  FnArgs extends GenericFunction['Args'],
-  PassedArgs extends GenericFunction['Args'],
-> = [FnArgs] extends [Record<PropertyKey, never>]
-  ? PassedArgs extends Record<PropertyKey, never>
-    ? true
-    : false
-  : keyof PassedArgs extends keyof FnArgs
-    ? PassedArgs extends FnArgs
-      ? true
-      : false
-    : false
-
-type MatchingFunctionArgs<
-  Fn extends GenericFunction,
-  Args extends GenericFunction['Args'],
-> = Fn extends { Args: infer A extends GenericFunction['Args'] }
-  ? IsMatchingArgs<A, Args> extends true
-    ? Fn
-    : never
+  ? [T] extends [never]
+    ? false
+    : true
   : false
-
-export type FindMatchingFunctionByArgs<
-  FnUnion,
-  Args extends GenericFunction['Args'],
-> = FnUnion extends infer Fn extends GenericFunction ? MatchingFunctionArgs<Fn, Args> : false
 
 type MatchingFunctionBySetofFrom<
   Fn extends GenericFunction,
-  TableName extends string,
+  TableName extends string
 > = Fn['SetofOptions'] extends GenericSetofOption
   ? TableName extends Fn['SetofOptions']['from']
     ? Fn
@@ -673,7 +635,7 @@ type MatchingFunctionBySetofFrom<
 
 type FindMatchingFunctionBySetofFrom<
   FnUnion,
-  TableName extends string,
+  TableName extends string
 > = FnUnion extends infer Fn extends GenericFunction
   ? MatchingFunctionBySetofFrom<Fn, TableName>
   : false
@@ -681,7 +643,7 @@ type FindMatchingFunctionBySetofFrom<
 type ComputedField<
   Schema extends GenericSchema,
   RelationName extends keyof TablesAndViews<Schema>,
-  FieldName extends keyof TablesAndViews<Schema>[RelationName]['Row'],
+  FieldName extends keyof TablesAndViews<Schema>[RelationName]['Row']
 > = FieldName extends keyof Schema['Functions']
   ? Schema['Functions'][FieldName] extends {
       Args: { '': TablesAndViews<Schema>[RelationName]['Row'] }
@@ -695,7 +657,7 @@ type ComputedField<
 // object, and the schema functions definitions
 export type GetComputedFields<
   Schema extends GenericSchema,
-  RelationName extends keyof TablesAndViews<Schema>,
+  RelationName extends keyof TablesAndViews<Schema>
 > = {
   [K in keyof TablesAndViews<Schema>[RelationName]['Row']]: ComputedField<Schema, RelationName, K>
 }[keyof TablesAndViews<Schema>[RelationName]['Row']]
