+++ packages/core/supabase-js/src/lib/rest/types/common/common.ts
@@ -1,0 +1,1 @@
+../../../../../../postgrest-js/src/types/common/common.ts
