+++ packages/core/supabase-js/src/lib/rest/types/common/rpc.ts
@@ -1,0 +1,1 @@
+../../../../../../postgrest-js/src/types/common/rpc.ts
